+++ src/Payment/Changelog.md
@@ -12,6 +12,7 @@
 - Added Laravel 11 Support
 - Changed minimum Laravel version to v10.38.2
 - Changed minimal Enum requirement to v4.2
+- BC: Added the `getRemoteId()` method to the `PaymentRequest` interface
 - BC: Changed the `PaymentMethod` interface into Configurable
 - Deprecated the `PaymentMethod::getConfiguration()` in favor of `configuration()`