+++ src/Payment/Changelog.md
@@ -21,7 +21,7 @@
 - BC: Added the `getTransactionAmount()` method to the `PaymentResponse` interface
 - BC: Changed the `PaymentMethod` interface into Configurable
 - Deprecated the `PaymentMethod::getConfiguration()` in favor of `configuration()`
-- Deprecated the `Payment::getAmountPaid()` method in favor of `getTransactionAmount()`
+- Deprecated the `PaymentResponse::getAmountPaid()` method in favor of `getTransactionAmount()`
 
 ## 3.x Series